feat: Restructure matching engine and plug it into app (#6)
