+++ ctenex/api/app_factory.py
@@ -1,19 +1,21 @@
 from contextlib import asynccontextmanager
+from typing import AsyncIterator
 
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from starlette.types import StatefulLifespan
 
 from ctenex import __version__
-from ctenex.api.v1.controllers.exchange import router as exchange_router
-from ctenex.api.v1.controllers.status import router as status_router
-from ctenex.domain.matching_engine import MatchingEngine
+from ctenex.api.controllers.status import router as status_router
+from ctenex.api.v1.controllers.in_memory.exchange import router as exchange_router
+from ctenex.domain.in_memory.matching_engine.model import MatchingEngine
 from ctenex.settings.application import get_app_settings
 
 settings = get_app_settings()
 
 
 @asynccontextmanager
-async def lifespan(app: FastAPI):
+async def lifespan(app: FastAPI) -> AsyncIterator:
     # Instantiate the matching engine
     app.state.matching_engine = MatchingEngine()
     app.state.matching_engine.start()
@@ -22,7 +24,7 @@ async def lifespan(app: FastAPI):
     app.state.matching_engine.stop()
 
 
-def create_app() -> FastAPI:
+def create_app(lifespan: StatefulLifespan[FastAPI] | None = None) -> FastAPI:
     app = FastAPI(
         title=settings.project_name,
         version=__version__,
